backend/requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
deep-translator==1.11.4
gtts==2.5.4
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
